requirements.txt
pymupdf
Pillow
easyocr
requests
numpy